+++ .github/workflows/smoke-test-prod.yml
@@ -13,7 +13,7 @@ jobs:
       fail-fast: false
       matrix:
         platform: [ubuntu-latest, macos-latest, windows-latest]
-        example: [react/gatsby-themed-site, react/create-react-app, react/next.js, vanillajs/monolith, vue/quasar-app, svelte/svelte-app]
+        example: [react/gatsby-themed-site, react/create-react-runtime, react/next.js, vanillajs/monolith, vue/quasar-runtime, svelte/svelte-runtime]
         include:
           - platform: ubuntu-latest
             CARGO_HOME: ~/.cargo
