+++ examples/multiwindow/src-tauri/tauri.conf.json
@@ -27,14 +27,12 @@
     "windows": [
       {
         "label": "Main",
-        "url": "app",
         "title": "Tauri - Main",
         "width": 800,
         "height": 600
       },
       {
         "label": "Secondary",
-        "url": "app",
         "title": "Tauri - Secondary",
         "width": 600,
         "height": 400
