+++ tauri-api/src/notification.rs
@@ -70,7 +70,7 @@ impl Notification {
       let exe = std::env::current_exe()?;
       let exe_dir = exe.parent().expect("failed to get exe directory");
       let curr_dir = exe_dir.display().to_string();
-      // set the notification's System.AppUserModel.ID only when running the installed app
+      // set the notification's System.AppUserModel.ID only when running the installed runtime
       if !(curr_dir.ends_with(format!("{S}target{S}debug", S = MAIN_SEPARATOR).as_str())
         || curr_dir.ends_with(format!("{S}target{S}release", S = MAIN_SEPARATOR).as_str()))
       {
