+++ tauri/src/endpoints.rs
@@ -1,4 +1,6 @@
-use crate::{api::config::Config, InvokeMessage, Manager};
+use crate::api::config::Config;
+use crate::hooks::InvokeMessage;
+use crate::runtime::Manager;
 use serde::{Deserialize, Serialize};
 use serde_json::Value as JsonValue;
